agent/pyproject.toml
[project]
name = "lenny-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.2",
    "openai>=1.12.0",
    "tiktoken>=0.5.0",
]
